Implementação inicial dos testes

+++ backendstoom/src/main/java/backendstoom/backendstoom/domain/Endereco.java
@@ -58,13 +58,13 @@ public class Endereco implements Serializable {
 	@NotNull(message = "Campo zipcode é obrigatório.")
 	private Long zipcode;
 	
-	private Long latitude;
-	private Long longitude;
+	private Double latitude;
+	private Double longitude;
 	
 	public Endereco () {}
 
 	public Endereco(String streetName, Long number, String complement, String neighbourhood, String city, String state,
-			String country, Long zipcode, Long latitude, Long longitude) {
+			String country, Long zipcode, Double latitude, Double longitude) {
 		this.streetName = streetName;
 		this.number = number;
 		this.complement = complement;
@@ -145,19 +145,19 @@ public class Endereco implements Serializable {
 		this.zipcode = zipcode;
 	}
 
-	public Long getLatitude() {
+	public Double getLatitude() {
 		return latitude;
 	}
 
-	public void setLatitude(Long latitude) {
+	public void setLatitude(Double latitude) {
 		this.latitude = latitude;
 	}
 
-	public Long getLongitude() {
+	public Double getLongitude() {
 		return longitude;
 	}
 
-	public void setLongitude(Long longitude) {
+	public void setLongitude(Double longitude) {
 		this.longitude = longitude;
 	}
